requirements.txt
asgiref==3.8.1
Django==5.1
sqlparse==0.5.1
django-cleanup==8.1.0
django-cors-headers==4.4.0
django-tinymce==4.1.0
django-unfold==0.34.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
packaging==24.1
pillow==10.4.0
PyJWT==2.9.0